requirements.txt
hyperliquid-python-sdk>=0.21.0
httpx>=0.27.0
pyyaml>=6.0.1
aiohttp>=3.9.0
